requirements.txt
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=0.24.0
sounddevice>=0.4.0
soundfile>=0.10.0
python_speech_features==0.6.1
opencv-python>=4.5.3.56
Pillow>=8.2.0
speechrecognition>=3.8.1
